Make Jedd with Polyglot process only the classes specified on the command line, but not fully analyze or output any context classes pulled in by references from the command line classes.
Also, allow both .jedd and .java extensions for source files. We first search
the entire sourcepath for a .jedd file. If there is none, we again search
the entire sourcepath from scratch for a .java file. Finally, we give up and
report an error.

Also, bug fixes to type checking of field and local references, and fix a crash
of the constraint generation when no constraints are generated.